Fix format specifier warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36483 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/monitor/mon_file.c
@@ -249,7 +249,7 @@ void mon_file_load(const char *filename, int device, MON_ADDR start_addr,
         curbank = origbank;
     }
 
-    mon_out("to %04X (%x bytes)\n", ADDR_LIMIT(adr + ch), ch);
+    mon_out("to %04X (%x bytes)\n", ADDR_LIMIT(adr + ch), (unsigned int)ch);
 
     /* set end of load addresses like kernal load if
      * 1. loading .prg file
